+++ src/Dialog/SynonymsDialog.cpp
@@ -37,6 +37,9 @@ SynonymsDialog::SynonymsDialog( QMap<QString, QVariant> word, QWidget *parent ) 
     QPushButton *saveButton = ui->buttonBox->button( QDialogButtonBox::Save );
     saveButton->setText( tr( "Save" ) );
     connect( saveButton, SIGNAL( clicked() ), this, SLOT( saveSynonyms() ) );
+
+    // May be there is other way to translate Core Strings
+    ui->buttonBox->button( QDialogButtonBox::Cancel )->setText( tr( "Cancel" ) );
 }
 
 SynonymsDialog::~SynonymsDialog()